refactor(Home): 페이지 별 컴포넌트 구분을 위한 수정
